Upload Breadcrumb report on control panel

+++ Sources/Features/Extensions/Date+Extensions.swift
@@ -1,11 +1,11 @@
 import Foundation
 
 extension Date {
-    var millisecondsSince1970:Int64 {
-        Int64((self.timeIntervalSince1970 * 1000.0).rounded())
+    var millisecondsSince1970:Int {
+        Int((self.timeIntervalSince1970 * 1000.0).rounded())
     }
     
-    init(milliseconds:Int64) {
+    init(milliseconds:Int) {
         self = Date(timeIntervalSince1970: TimeInterval(milliseconds) / 1000)
     }
 }
